rename Publisher#promise() to #resolve(), add Publisher#reject()
Just like the usual `#resolve()` method of a promise which is to be called with a successful result or another promise object of which the result is then used, `Publisher` can now handle both scenarios as well. The method is now also complemented by the `#reject()` method for resolving the action as a failure.